Fix initial selection when value is not valid
When the `value` prop is set, the option with the corresponding value will be selected. Previously, if the value was not present in the options list, the first option was automatically selected. With this change, if an invalid value is set, none of the options will be selected, and the `value` prop will return an empty string.